pjax_modal -> ajax_modal_rails
