+++ app/controllers/match_client_details_controller.rb
@@ -6,7 +6,7 @@
 
 class MatchClientDetailsController < ApplicationController
   include HasMatchAccessContext
-  include PjaxModalController
+  include AjaxModalRails::Controller
 
   skip_before_action :authenticate_user!
   before_action :require_match_access_context!
