+++ app/views/ajax_modal_rails/_content.haml
@@ -1,0 +1,25 @@
+:ruby
+  title = yield :modal_title
+  subtitle = yield :modal_title_secondary
+
+.modal-header
+  .modal-header__titles
+    - if title.present?
+      %h4.modal-title(data-ajax-modal-title)
+        = title
+    - if subtitle.present?
+      %h5.modal-title--secondary(data-ajax-modal-title-secondary)
+        = subtitle
+  %button.close(aria-label="Close" ajax-modal-close type="button")
+    %span.icon-cross(aria-hidden="true")
+.modal-body.clearfix
+  %div{"data-ajax-modal-body" => ""}
+    = yield
+  %div{"data-ajax-modal-loading" => "", :style => "display: none;"}
+    Loading&hellip;
+- if content_for?(:modal_footer) || content_for?(:modal_only_footer)
+  .modal-footer{"data-ajax-modal-footer" => ""}
+    = yield :modal_footer
+    = yield :modal_only_footer
+
+= yield :page_js
