+++ app/views/ajax_modal_rails/_frame.html.haml
@@ -1,0 +1,16 @@
+#ajax-modal.modal.fade{"data-ajax-modal" => "", :role => "dialog", :tabindex => "-1", aria_hidden: true}
+  .modal-dialog.modal-dialog-scrollable.modal-dialog-centered{:role => "document"}
+    .modal-content{"data-ajax-modal-content" => ""}
+      .modal-header
+        %h4.modal-title{"data-ajax-modal-title" => ""} Loading
+        %button.close(aria-label="Close" data-ajax-modal-close type="button")
+          %span.icon-cross(aria-hidden="true")
+      .modal-body.clearfix
+        %div{"data-ajax-modal-loading" => ""}
+          Loading&hellip;
+
+
+- content_for :page_js do
+  :javascript
+    window.ajaxModalInstance = new AjaxModal()
+
