+++ app/views/ajax_modal_rails/close_modal_and_reload_page.html.erb
@@ -1,0 +1,3 @@
+<script>
+  window.ajaxModalInstance.closeAndReload()
+</script>
