+++ app/views/ajax_modal_rails/redirect_via_js.html.erb
@@ -1,0 +1,4 @@
+<script>
+  window.location.href = '<%= redirect %>'
+</script>
+
