--- app/views/application/_pjax_modal_frame.haml
@@ -1,10 +1,0 @@
-#pjax-modal.modal.fade(data-pjax-modal role="dialog")
-  .modal-dialog(role="document")
-    .modal-content(data-pjax-modal-container)
-      .modal-header
-        %h4.modal-title(data-pjax-modal-title)
-        %button.close(aria-label="Close" pjax-modal-close type="button")
-          %span.icon-cross(aria-hidden="true")
-      .modal-body.clearfix
-        %div(data-pjax-modal-loading)
-          Loading&hellip;
