+++ app/views/help/create.js.coffee
@@ -2,7 +2,7 @@
   $('.help-form').html "<%=j render 'form' %>"
   $('.jSwitcher').trigger('change');
 <% else %>
-  #window.pjax_modal_instance.closeModal()
-  $('#pjax-modal').modal('hide')
+  #window.ajaxModalInstance.closeModal()
+  $('#ajax-modal').modal('hide')
   #window.location.reload()
 <% end %>
