--- app/views/layouts/pjax_modal_content.haml
@@ -1,15 +1,0 @@
-.modal-header
-  %h4.modal-title(data-pjax-modal-title)
-    = yield :modal_title
-  %button.close(aria-label="Close" pjax-modal-close type="button")
-    %span.icon-cross(aria-hidden="true")
-.modal-body.clearfix
-  %div(data-pjax-modal-body)
-    = yield
-  %div(data-pjax-modal-loading style="display: none;")
-    Loading&hellip;
-- if content_for?(:modal_footer) || content_for?(:modal_only_footer)
-  .modal-footer(data-pjax-modal-footer)
-    = yield :modal_footer
-    = yield :modal_only_footer
-= yield :page_js
