+++ app/views/match_contacts/_form.haml
@@ -1,4 +1,4 @@
-= simple_form_for @match_contacts, url: access_context.match_contacts_path(@match), html: { class: 'current-step-form jContactForm' }, remote: pjax_request? || request.xhr? do |form|
+= simple_form_for @match_contacts, url: access_context.match_contacts_path(@match), html: { class: 'current-step-form jContactForm' }, remote: ajax_modal_request? || request.xhr? do |form|
   = render 'contact_manager/contact_field', contact_type: :shelter_agency_contacts, input_name: 'match_contacts[shelter_agency_contact_ids][]', object: @match_contacts, form: form, current_contact: @current_contact
 
   = render 'contact_manager/contact_field', contact_type: :client_contacts, input_name: 'match_contacts[client_contact_ids][]', object: @match_contacts, form: form, current_contact: @current_contact
